refactor(structure): restructure code so we can properly support many algorithms
