+++ causy/graph.py
@@ -1,27 +1,15 @@
-from abc import ABC
 from dataclasses import dataclass
 from typing import List, Optional, Dict, Set, Tuple, Union
 from uuid import uuid4
 import logging
 
 import torch
-import torch.multiprocessing as mp
 
-from causy.interfaces import (
-    IndependenceTestInterface,
-)
 from causy.interfaces import (
     BaseGraphInterface,
     NodeInterface,
     TestResultAction,
     TestResult,
-    GraphModelInterface,
-    LogicStepInterface,
-    ExitConditionInterface,
-)
-from causy.utils import (
-    load_pipeline_artefact_by_definition,
-    load_pipeline_steps_by_definition,
 )
 
 logger = logging.getLogger(__name__)
@@ -422,279 +410,3 @@ class Graph(BaseGraphInterface):
                     # TODO: check if this is correct (@sof)
                     return True
         return False
-
-
-def unpack_run(args):
-    tst = args[0]
-    del args[0]
-    return tst(*args)
-
-
-class AbstractGraphModel(GraphModelInterface, ABC):
-    """
-    The graph model is the main class of causy. It is responsible for creating a graph from data and executing the pipeline_steps.
-
-    The graph model is responsible for the following tasks:
-    - Create a graph from data (create_graph_from_data)
-    - Execute the pipeline_steps (execute_pipeline_steps)
-    - Take actions on the graph (execute_pipeline_step & _take_action which is called by execute_pipeline_step)
-
-    It also initializes and takes care of the multiprocessing pool.
-
-    """
-
-    pipeline_steps: List[IndependenceTestInterface]
-    graph: BaseGraphInterface
-    pool: mp.Pool
-
-    def __init__(
-        self,
-        graph=None,
-        pipeline_steps: Optional[List[IndependenceTestInterface]] = None,
-    ):
-        self.graph = graph
-        self.pipeline_steps = pipeline_steps or []
-        self.pool = mp.Pool(mp.cpu_count() * 2)
-
-    def create_graph_from_data(self, data: List[Dict[str, float]]):
-        """
-        Create a graph from data
-        :param data: is a list of dictionaries
-        :return:
-        """
-        # initialize nodes
-        keys = data[0].keys()
-        nodes: Dict[str, List[float]] = {}
-
-        for key in keys:
-            nodes[key] = []
-
-        # load nodes into node dict
-        for row in data:
-            for key in keys:
-                nodes[key].append(row[key])
-
-        graph = Graph()
-        for key in keys:
-            graph.add_node(key, nodes[key])
-
-        self.graph = graph
-        return graph
-
-    def create_all_possible_edges(self):
-        """
-        Create all possible edges on a graph
-        TODO: replace me with the skeleton builders
-        :return:
-        """
-        for u in self.graph.nodes.values():
-            for v in self.graph.nodes.values():
-                if u == v:
-                    continue
-                self.graph.add_edge(u, v, {})
-
-    def execute_pipeline_steps(self):
-        """
-        Execute all pipeline_steps
-        :return: the steps taken during the step execution
-        """
-        action_history = []
-
-        for filter in self.pipeline_steps:
-            logger.info(f"Executing pipeline step {filter.__class__.__name__}")
-            if isinstance(filter, LogicStepInterface):
-                filter.execute(self.graph, self)
-                continue
-
-            result = self.execute_pipeline_step(filter)
-            action_history.append(
-                {"step": filter.__class__.__name__, "actions": result}
-            )
-
-        return action_history
-
-    def _format_yield(self, test_fn, graph, generator):
-        """
-        Format the yield for the parallel processing
-        :param test_fn: the pipeline_step test function
-        :param graph: the graph
-        :param generator: the generator object which generates the combinations
-        :return: yields the test function with its inputs
-        """
-        for i in generator:
-            yield [test_fn, [*i], graph]
-
-    def _take_action(self, results):
-        """
-        Take the actions returned by the test
-
-        In causy changes on the graph are not executed directly. Instead, the test returns an action which should be executed on the graph.
-        This is done to make it possible to execute the tests in parallel as well as to decide proactively at which point in the decisions taken by the pipeline step should be executed.
-        Actions are returned by the test and are executed on the graph. The actions are stored in the action history to make it possible to revert the actions or use them in a later step.
-
-        :param results:
-        :return:
-        """
-        actions_taken = []
-        for result_items in results:
-            if result_items is None:
-                continue
-            if not isinstance(result_items, list):
-                result_items = [result_items]
-
-            for i in result_items:
-                if i.x is not None and i.y is not None:
-                    logger.info(f"Action: {i.action} on {i.x.name} and {i.y.name}")
-
-                # execute the action returned by the test
-                if i.action == TestResultAction.REMOVE_EDGE_UNDIRECTED:
-                    if not self.graph.undirected_edge_exists(i.x, i.y):
-                        logger.debug(
-                            f"Tried to remove undirected edge {i.x.name} <-> {i.y.name}. But it does not exist."
-                        )
-                        continue
-                    self.graph.remove_edge(i.x, i.y)
-                    self.graph.add_edge_history(i.x, i.y, i)
-                    self.graph.add_edge_history(i.y, i.x, i)
-                elif i.action == TestResultAction.UPDATE_EDGE:
-                    if not self.graph.edge_exists(i.x, i.y):
-                        logger.debug(
-                            f"Tried to update edge {i.x.name} -> {i.y.name}. But it does not exist."
-                        )
-                        continue
-                    self.graph.update_edge(i.x, i.y, i.data)
-                    self.graph.add_edge_history(i.x, i.y, i)
-                    self.graph.add_edge_history(i.y, i.x, i)
-                elif i.action == TestResultAction.UPDATE_EDGE_DIRECTED:
-                    if not self.graph.directed_edge_exists(i.x, i.y):
-                        logger.debug(
-                            f"Tried to update directed edge {i.x.name} -> {i.y.name}. But it does not exist."
-                        )
-                        continue
-                    self.graph.update_directed_edge(i.x, i.y, i.data)
-                    self.graph.add_edge_history(i.x, i.y, i)
-                elif i.action == TestResultAction.DO_NOTHING:
-                    continue
-                elif i.action == TestResultAction.REMOVE_EDGE_DIRECTED:
-                    if not self.graph.directed_edge_exists(i.x, i.y):
-                        logger.debug(
-                            f"Tried to remove directed edge {i.x.name} -> {i.y.name}. But it does not exist."
-                        )
-                        continue
-                    self.graph.remove_directed_edge(i.x, i.y)
-                    self.graph.add_edge_history(i.x, i.y, i)
-
-                # add the action to the actions history
-                actions_taken.append(i)
-
-        return actions_taken
-
-    def execute_pipeline_step(self, test_fn: IndependenceTestInterface):
-        """
-        Execute a single pipeline_step on the graph. either in parallel or in a single process depending on the test_fn.parallel flag
-        :param test_fn: the test function
-        :param threshold: the threshold
-        :return:
-        """
-        actions_taken = []
-
-        # initialize the worker pool (we currently use all available cores * 2)
-
-        # run all combinations in parallel except if the number of combinations is smaller then the chunk size
-        # because then we would create more overhead then we would definetly gain from parallel processing
-        if test_fn.parallel:
-            for result in self.pool.imap_unordered(
-                unpack_run,
-                self._format_yield(
-                    test_fn, self.graph, test_fn.generator.generate(self.graph, self)
-                ),
-                chunksize=test_fn.chunk_size_parallel_processing,
-            ):
-                if not isinstance(result, list):
-                    result = [result]
-                actions_taken.extend(self._take_action(result))
-        else:
-            if test_fn.generator.chunked:
-                for chunk in test_fn.generator.generate(self.graph, self):
-                    iterator = [
-                        unpack_run(i)
-                        for i in [[test_fn, [*c], self.graph] for c in chunk]
-                    ]
-                    actions_taken.extend(self._take_action(iterator))
-            else:
-                iterator = [
-                    unpack_run(i)
-                    for i in [
-                        [test_fn, [*i], self.graph]
-                        for i in test_fn.generator.generate(self.graph, self)
-                    ]
-                ]
-                actions_taken.extend(self._take_action(iterator))
-
-        self.graph.action_history.append(
-            {"step": type(test_fn).__name__, "actions": actions_taken}
-        )
-
-        return actions_taken
-
-
-def graph_model_factory(
-    pipeline_steps: Optional[List[IndependenceTestInterface]] = None,
-) -> type[AbstractGraphModel]:
-    """
-    Create a graph model based on a List of pipeline_steps
-    :param pipeline_steps: a list of pipeline_steps which should be applied to the graph
-    :return: the graph model
-    """
-
-    class GraphModel(AbstractGraphModel):
-        def __init__(self):
-            super().__init__(pipeline_steps=pipeline_steps)
-
-    return GraphModel
-
-
-class Loop(LogicStepInterface):
-    """
-    A loop which executes a list of pipeline_steps until the exit_condition is met.
-    """
-
-    def execute(
-        self, graph: BaseGraphInterface, graph_model_instance_: GraphModelInterface
-    ):
-        """
-        Executes the loop til self.exit_condition is met
-        :param graph:
-        :param graph_model_instance_:
-        :return:
-        """
-        n = 0
-        steps = None
-        while not self.exit_condition(
-            graph=graph,
-            graph_model_instance_=graph_model_instance_,
-            actions_taken=steps,
-            iteration=n,
-        ):
-            steps = []
-            for pipeline_step in self.pipeline_steps:
-                result = graph_model_instance_.execute_pipeline_step(pipeline_step)
-                steps.extend(result)
-            n += 1
-
-    def __init__(
-        self,
-        pipeline_steps: Optional[List[IndependenceTestInterface]] = None,
-        exit_condition: ExitConditionInterface = None,
-    ):
-        super().__init__()
-        # TODO check if this is a good idea
-        if isinstance(exit_condition, dict):
-            exit_condition = load_pipeline_artefact_by_definition(exit_condition)
-
-        # TODO: check if this is a good idea
-        if len(pipeline_steps) > 0 and isinstance(pipeline_steps[0], dict):
-            pipeline_steps = load_pipeline_steps_by_definition(pipeline_steps)
-
-        self.pipeline_steps = pipeline_steps or []
-        self.exit_condition = exit_condition
