+++ tests/test_orientation_tests.py
@@ -1,15 +1,15 @@
 import unittest
 
-from causy.graph import graph_model_factory, Graph
+from causy.graph import Graph
 from causy.interfaces import TestResult, TestResultAction
-from causy.orientation_tests import (
+from causy.orientation_rules.pc import (
     ColliderTest,
     NonColliderTest,
     FurtherOrientTripleTest,
     OrientQuadrupleTest,
     FurtherOrientQuadrupleTest,
 )
-from causy.utils import retrieve_edges
+from causy.graph_model import graph_model_factory
 
 
 class OrientationRuleTestCase(unittest.TestCase):
